requirements.txt
fastapi
uvicorn
transformers
pandas
streamlit
python-dotenv